added router.js, moved app.js crud to tareas.js and app.js has components of login, register and tareas
