+++ client/src/routes/routes.js
@@ -1,0 +1,8 @@
+const routes = {
+    login: '/login',
+    register: '/register',
+    tareas: '/tareas',
+  };
+  
+  export default routes;
+
